Update tests/regular_sml_baseline.svg to match latest picosvg output
The write_symbol_from_svgs.py test fails with the latest picosvg package, the expected svg needs to be updated to match the current output. The difference is minor and only in the text diff, visually it appears the same as before

I need this so that I can upgrade picosvg in google3 to the latest version, otherwise TAP for vf2symbols fails with this spurious test failure.